+++ index.html
@@ -229,6 +229,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-<script src="main.js"></script>
+<script src="main.js" type="module"></script>
 </body>
 </html>
